adding search

+++ ZivotopisIN2/src/main/java/com/DiplomskiRad_SK/ZivotopisIN2/services/ZivotopisDBService.java
@@ -8,12 +8,15 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.DiplomskiRad_SK.ZivotopisIN2.controller.UploadController;
 import com.DiplomskiRad_SK.ZivotopisIN2.modelDB.CV;
 import com.DiplomskiRad_SK.ZivotopisIN2.modelDB.DodatneInfo;
 import com.DiplomskiRad_SK.ZivotopisIN2.modelDB.Drzava;
@@ -59,6 +62,7 @@ import oracle.net.aso.o;
 
 @Service("BLCV")
 public class ZivotopisDBService {
+	private static final Logger log = LogManager.getLogger(UploadController.class);
 
 	@Autowired
 	CVRepository cvRepo;
@@ -107,215 +111,223 @@ public class ZivotopisDBService {
 	@Autowired
 	ZnaRepository znaRepo;
 
-
-
 	@Transactional
 	public void SaveCV(CV cv) {
 
 		cv = CheckForDuplicatesInDB(cv);
-		
 		cvRepo.save(cv);
-		// cv.getOsobnaVjestinaList().get(1).getVozackaDozvolaList().get(1).getIdOsobnaVj()
-		/*
-		 * CV cv1 = new CV(); Iterable<EdukacijaITrening> a =
-		 * cv1.getEdukacijaITreningList(); ArrayList<CV> b = new ArrayList<CV>();
-		 * EdukacijaITrening[] c = (EdukacijaITrening[]) ((Collection<CV>)
-		 * a).toArray(new EdukacijaITrening[a.size()])
-		 */
 	}
 
 	private CV CheckForDuplicatesInDB(CV cv) {
-		// Dodatne info
-				for (int i = 0; i < cv.getDodatneInfoList().size(); i++) {
-					Kategorija katDB = kategorijaRepo.findByNaziv(cv.getDodatneInfoList().get(i).getKategorija().getNaziv());
-					if (katDB == null) {
-						katDB = cv.getDodatneInfoList().get(i).getKategorija();
-						katDB.setDependenciesNull();
-						kategorijaRepo.save(katDB);
-					}
-					cv.getDodatneInfoList().get(i).setKategorija(katDB);
-
+		try {
+			// Dodatne info
+			for (int i = 0; i < cv.getDodatneInfoList().size(); i++) {
+				Kategorija katDB = kategorijaRepo
+						.findByNaziv(cv.getDodatneInfoList().get(i).getKategorija().getNaziv());
+				if (katDB == null) {
+					katDB = cv.getDodatneInfoList().get(i).getKategorija();
+					katDB.setDependenciesNull();
+					kategorijaRepo.save(katDB);
 				}
+				cv.getDodatneInfoList().get(i).setKategorija(katDB);
 
-				// Vozacka dozvola
-				for (int i = 0; i < cv.getOsobnaVJ().getVozackaDozvolaOsVJList().size(); i++) {
-					VozackaDozvola vozDozDB = vozackaDozRepo.findByKategorija(
-							cv.getOsobnaVJ().getVozackaDozvolaOsVJList().get(i).getVozackaDozvola().getKategorija());
-					if (vozDozDB == null) {
-						vozDozDB = cv.getOsobnaVJ().getVozackaDozvolaOsVJList().get(i).getVozackaDozvola();
-						vozDozDB.setDependenciesNull();
-						vozackaDozRepo.save(vozDozDB);
-					}
-					cv.getOsobnaVJ().getVozackaDozvolaOsVJList().get(i).setVozackaDozvola(vozDozDB);
+			}
 
+			// Vozacka dozvola
+			for (int i = 0; i < cv.getOsobnaVJ().getVozackaDozvolaOsVJList().size(); i++) {
+				VozackaDozvola vozDozDB = vozackaDozRepo.findByKategorija(
+						cv.getOsobnaVJ().getVozackaDozvolaOsVJList().get(i).getVozackaDozvola().getKategorija());
+				if (vozDozDB == null) {
+					vozDozDB = cv.getOsobnaVJ().getVozackaDozvolaOsVJList().get(i).getVozackaDozvola();
+					vozDozDB.setDependenciesNull();
+					vozackaDozRepo.save(vozDozDB);
 				}
+				cv.getOsobnaVJ().getVozackaDozvolaOsVJList().get(i).setVozackaDozvola(vozDozDB);
 
-				// Jezik
-				for (int i = 0; i < cv.getOsobnaVJ().getZnaList().size(); i++) {
-					Jezik jezikDB = jezikRepo.findByNaziv(cv.getOsobnaVJ().getZnaList().get(i).getJezik().getNaziv());
-					if (jezikDB == null) {
-						jezikDB = cv.getOsobnaVJ().getZnaList().get(i).getJezik();
-						jezikDB.setDependenciesNull();
-						jezikRepo.save(jezikDB);
-					}
-					cv.getOsobnaVJ().getZnaList().get(i).setJezik(jezikDB);
+			}
 
+			// Jezik
+			for (int i = 0; i < cv.getOsobnaVJ().getZnaList().size(); i++) {
+				Jezik jezikDB = jezikRepo.findByNaziv(cv.getOsobnaVJ().getZnaList().get(i).getJezik().getNaziv());
+				if (jezikDB == null) {
+					jezikDB = cv.getOsobnaVJ().getZnaList().get(i).getJezik();
+					jezikDB.setDependenciesNull();
+					jezikRepo.save(jezikDB);
 				}
+				cv.getOsobnaVJ().getZnaList().get(i).setJezik(jezikDB);
 
-				// Osoba
-				ArrayList<Osoba> osobaDBList = (ArrayList<Osoba>) osobaRepo.findAll();
-				for (Osoba osobaDB : osobaDBList) {
-					if (cv.getOsoba().equals(osobaDB)) {
-						cv.setOsoba(osobaDB);
+			}
 
-						for (int i = 0; i < cv.getOsoba().getKontaktInfoList().size(); i++)
-							cv.getOsoba().getKontaktInfoList().get(i).setOsoba(osobaDB);
+			// Osoba
+			ArrayList<Osoba> osobaDBList = (ArrayList<Osoba>) osobaRepo.findAll();
+			for (Osoba osobaDB : osobaDBList) {
+				if (cv.getOsoba().equals(osobaDB)) {
+					cv.setOsoba(osobaDB);
 
-						for (int i = 0; i < cv.getOsoba().getOsobaDrzavljanstvoList().size(); i++)
-							cv.getOsoba().getOsobaDrzavljanstvoList().get(i).setOsoba(osobaDB);
-					}
+					for (int i = 0; i < cv.getOsoba().getKontaktInfoList().size(); i++)
+						cv.getOsoba().getKontaktInfoList().get(i).setOsoba(osobaDB);
+
+					for (int i = 0; i < cv.getOsoba().getOsobaDrzavljanstvoList().size(); i++)
+						cv.getOsoba().getOsobaDrzavljanstvoList().get(i).setOsoba(osobaDB);
 				}
+			}
 
-				// Tip Konatakta
-				for (int i = 0; i < cv.getOsoba().getKontaktInfoList().size(); i++) {
-					TipKontakta tipKontDB = tipKontRepo
-							.findByNaziv(cv.getOsoba().getKontaktInfoList().get(i).getTipKontakta().getNaziv());
-					if (tipKontDB == null) {
-						tipKontDB = cv.getOsoba().getKontaktInfoList().get(i).getTipKontakta();
-						tipKontDB.setDependenciesNull();
-						tipKontRepo.save(tipKontDB);
-					}
-					cv.getOsoba().getKontaktInfoList().get(i).setTipKontakta(tipKontDB);
-
+			// Tip Konatakta
+			for (int i = 0; i < cv.getOsoba().getKontaktInfoList().size(); i++) {
+				TipKontakta tipKontDB = tipKontRepo
+						.findByNaziv(cv.getOsoba().getKontaktInfoList().get(i).getTipKontakta().getNaziv());
+				if (tipKontDB == null) {
+					tipKontDB = cv.getOsoba().getKontaktInfoList().get(i).getTipKontakta();
+					tipKontDB.setDependenciesNull();
+					tipKontRepo.save(tipKontDB);
 				}
+				cv.getOsoba().getKontaktInfoList().get(i).setTipKontakta(tipKontDB);
 
-				// Drzavljanstvo
-				for (int i = 0; i < cv.getOsoba().getOsobaDrzavljanstvoList().size(); i++) {
-					Drzavljanstvo drzavljanstvoDB = drzavljanstvoRepo
-							.findByNaziv(cv.getOsoba().getOsobaDrzavljanstvoList().get(i).getDrzavljanstvo().getNaziv());
-					if (drzavljanstvoDB == null)  {
-						drzavljanstvoDB = cv.getOsoba().getOsobaDrzavljanstvoList().get(i).getDrzavljanstvo();
-						drzavljanstvoDB.setDependenciesNull();
-						drzavljanstvoRepo.save(drzavljanstvoDB); 
-					}
-					cv.getOsoba().getOsobaDrzavljanstvoList().get(i).setDrzavljanstvo(drzavljanstvoDB);
+			}
 
+			// Drzavljanstvo
+			for (int i = 0; i < cv.getOsoba().getOsobaDrzavljanstvoList().size(); i++) {
+				Drzavljanstvo drzavljanstvoDB = drzavljanstvoRepo
+						.findByNaziv(cv.getOsoba().getOsobaDrzavljanstvoList().get(i).getDrzavljanstvo().getNaziv());
+				if (drzavljanstvoDB == null) {
+					drzavljanstvoDB = cv.getOsoba().getOsobaDrzavljanstvoList().get(i).getDrzavljanstvo();
+					drzavljanstvoDB.setDependenciesNull();
+					drzavljanstvoRepo.save(drzavljanstvoDB);
 				}
+				cv.getOsoba().getOsobaDrzavljanstvoList().get(i).setDrzavljanstvo(drzavljanstvoDB);
 
-				// VrstaPrijave
-				VrstaPrijave vrstaPDB = vrstaPrijRepo.findByNaziv(cv.getZaglavlje().getVrstaPrijave().getNaziv());
-				if (vrstaPDB == null) {
-					vrstaPDB = cv.getZaglavlje().getVrstaPrijave();
-					vrstaPDB.setDependenciesNull();
-					vrstaPrijRepo.save(vrstaPDB);
+			}
+
+			// VrstaPrijave
+			VrstaPrijave vrstaPDB = vrstaPrijRepo.findByNaziv(cv.getZaglavlje().getVrstaPrijave().getNaziv());
+			if (vrstaPDB == null) {
+				vrstaPDB = cv.getZaglavlje().getVrstaPrijave();
+				vrstaPDB.setDependenciesNull();
+				vrstaPrijRepo.save(vrstaPDB);
+			}
+			cv.getZaglavlje().setVrstaPrijave(vrstaPDB);
+
+			// Pozicija
+			for (int i = 0; i < cv.getRadnoIskustvoList().size(); i++) {
+				Pozicija pozicijaDB = pozicijaRepo
+						.findByNaziv(cv.getRadnoIskustvoList().get(i).getPozicija().getNaziv());
+				if (pozicijaDB == null) {
+					pozicijaDB = cv.getRadnoIskustvoList().get(i).getPozicija();
+					pozicijaDB.setDependenciesNull();
+					pozicijaRepo.save(pozicijaDB);
 				}
-				cv.getZaglavlje().setVrstaPrijave(vrstaPDB);
+				cv.getRadnoIskustvoList().get(i).setPozicija(pozicijaDB);
 
+			}
 
-				// Pozicija
-				for (int i = 0; i < cv.getRadnoIskustvoList().size(); i++) {
-					Pozicija pozicijaDB = pozicijaRepo.findByNaziv(cv.getRadnoIskustvoList().get(i).getPozicija().getNaziv());
-					if (pozicijaDB == null) {
-						pozicijaDB = cv.getRadnoIskustvoList().get(i).getPozicija();
-						pozicijaDB.setDependenciesNull();
-						pozicijaRepo.save(pozicijaDB);
-					}
-					cv.getRadnoIskustvoList().get(i).setPozicija(pozicijaDB);
+			// Drzava - osoba
+			Drzava drzavaDB = drzavaRepo.findByNaziv(cv.getOsoba().getMjesto().getDrzava().getNaziv());
+			if (drzavaDB == null) {
+				drzavaDB = cv.getOsoba().getMjesto().getDrzava();
+				drzavaDB.setDependenciesNull();
+				drzavaRepo.save(drzavaDB);
+			}
+			cv.getOsoba().getMjesto().setDrzava(drzavaDB);
 
+			// Mjesto - osoba
+			Mjesto mjestoDB = mjestoRepo.findByPBR(cv.getOsoba().getMjesto().getPBR());
+			if (mjestoDB == null) {
+				mjestoDB = cv.getOsoba().getMjesto();
+				mjestoDB.setDependenciesNull();
+				mjestoDB.setDrzava(drzavaDB);
+				mjestoRepo.save(mjestoDB);
+			}
+			cv.getOsoba().setMjesto(mjestoDB);
+
+			// Mjesto - Drzava RI
+			for (int i = 0; i < cv.getRadnoIskustvoList().size(); i++) {
+				Mjesto mjestoDB_RI = mjestoRepo
+						.findByPBR(cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto().getPBR());
+				Drzava drzavaDB_RI = drzavaRepo.findByNaziv(
+						cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto().getDrzava().getNaziv());
+
+				if (drzavaDB_RI == null) {
+					drzavaDB_RI = cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto().getDrzava();
+					drzavaDB_RI.setDependenciesNull();
+					drzavaRepo.save(drzavaDB_RI);
 				}
+				cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto().setDrzava(drzavaDB_RI);
 
-				//Drzava - osoba
-				Drzava drzavaDB = drzavaRepo.findByNaziv(cv.getOsoba().getMjesto().getDrzava().getNaziv());
-				if (drzavaDB == null) {
-					drzavaDB = cv.getOsoba().getMjesto().getDrzava();
-					drzavaDB.setDependenciesNull();
-					drzavaRepo.save(drzavaDB);
+				if (mjestoDB_RI == null) {
+					mjestoDB_RI = cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto();
+					mjestoDB_RI.setDependenciesNull();
+					mjestoDB_RI.setDrzava(drzavaDB_RI);
+					mjestoRepo.save(mjestoDB_RI);
 				}
-				cv.getOsoba().getMjesto().setDrzava(drzavaDB);
+				cv.getRadnoIskustvoList().get(i).getInstitucija().setMjesto(mjestoDB_RI);
 
-				//Mjesto - osoba
-				Mjesto mjestoDB = mjestoRepo.findByPBR(cv.getOsoba().getMjesto().getPBR());
-				if (mjestoDB == null) {
-					mjestoDB = cv.getOsoba().getMjesto();
-					mjestoDB.setDependenciesNull();
-					mjestoDB.setDrzava(drzavaDB);
-					mjestoRepo.save(mjestoDB);
+			}
+
+			// Mjesto - Drzava Edu
+			for (int i = 0; i < cv.getEdukacijaITreningList().size(); i++) {
+				Mjesto mjestoDB_ET = mjestoRepo
+						.findByPBR(cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto().getPBR());
+				Drzava drzavaDB_ET = drzavaRepo.findByNaziv(
+						cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto().getDrzava().getNaziv());
+
+				if (drzavaDB_ET == null) {
+					drzavaDB_ET = cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto().getDrzava();
+					drzavaDB_ET.setDependenciesNull();
+					drzavaRepo.save(drzavaDB_ET);
 				}
-				cv.getOsoba().setMjesto(mjestoDB);
+				cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto().setDrzava(drzavaDB_ET);
 
-				
-				//Mjesto - Drzava RI
-				for (int i = 0; i < cv.getRadnoIskustvoList().size(); i++) {
-					Mjesto mjestoDB_RI = mjestoRepo.findByPBR(cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto().getPBR());
-					Drzava drzavaDB_RI = drzavaRepo.findByNaziv(cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto().getDrzava().getNaziv());
-					
-					if (drzavaDB_RI == null) {
-						drzavaDB_RI = cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto().getDrzava();
-						drzavaDB_RI.setDependenciesNull();
-						drzavaRepo.save(drzavaDB_RI);
-					}
-					cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto().setDrzava(drzavaDB_RI);
-
-					if (mjestoDB_RI == null) {
-						mjestoDB_RI = cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto();
-						mjestoDB_RI.setDependenciesNull();
-						mjestoDB_RI.setDrzava(drzavaDB_RI);
-						mjestoRepo.save(mjestoDB_RI);
-					}
-					cv.getRadnoIskustvoList().get(i).getInstitucija().setMjesto(mjestoDB_RI);
-
+				if (mjestoDB_ET == null) {
+					mjestoDB_ET = cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto();
+					mjestoDB_ET.setDependenciesNull();
+					mjestoDB_ET.setDrzava(drzavaDB_ET);
+					mjestoRepo.save(mjestoDB_ET);
 				}
-				
-				//Mjesto - Drzava Edu
-				for (int i = 0; i < cv.getEdukacijaITreningList().size(); i++) {
-					Mjesto mjestoDB_ET = mjestoRepo.findByPBR(cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto().getPBR());
-					Drzava drzavaDB_ET = drzavaRepo.findByNaziv(cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto().getDrzava().getNaziv());
-					
-					if (drzavaDB_ET == null) {
-						drzavaDB_ET = cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto().getDrzava();
-						drzavaDB_ET.setDependenciesNull();
-						drzavaRepo.save(drzavaDB_ET);
-					}
-					cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto().setDrzava(drzavaDB_ET);
+				cv.getEdukacijaITreningList().get(i).getInstitucija().setMjesto(mjestoDB_ET);
 
-					if (mjestoDB_ET == null) {
-						mjestoDB_ET = cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto();
-						mjestoDB_ET.setDependenciesNull();
-						mjestoDB_ET.setDrzava(drzavaDB_ET);
-						mjestoRepo.save(mjestoDB_ET);
-					}
-					cv.getEdukacijaITreningList().get(i).getInstitucija().setMjesto(mjestoDB_ET);
+			}
 
+			// Institucija //firme s istima nazivom?
+			for (int i = 0; i < cv.getRadnoIskustvoList().size(); i++) {
+				Institucija institucijaDB = institucijaRepo
+						.findByNaziv(cv.getRadnoIskustvoList().get(i).getInstitucija().getNaziv());
+				if (institucijaDB == null) {
+					institucijaDB = cv.getRadnoIskustvoList().get(i).getInstitucija();
+					institucijaDB.setDependenciesNull();
+					institucijaDB.setMjesto(cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto());
+					institucijaRepo.save(institucijaDB);
 				}
-				
-				//Institucija //firme s istima nazivom?
-				for (int i = 0; i < cv.getRadnoIskustvoList().size(); i++) {
-					Institucija institucijaDB = institucijaRepo.findByNaziv(cv.getRadnoIskustvoList().get(i).getInstitucija().getNaziv());
-					if (institucijaDB == null) {
-						institucijaDB = cv.getRadnoIskustvoList().get(i).getInstitucija();
-						institucijaDB.setDependenciesNull();
-						institucijaDB.setMjesto(cv.getRadnoIskustvoList().get(i).getInstitucija().getMjesto());
-						institucijaRepo.save(institucijaDB);
-					}
-					cv.getRadnoIskustvoList().get(i).setInstitucija(institucijaDB);
+				cv.getRadnoIskustvoList().get(i).setInstitucija(institucijaDB);
 
-				}
-				for (int i = 0; i < cv.getEdukacijaITreningList().size(); i++) {
-					Institucija institucijaDB = institucijaRepo.findByNaziv(cv.getEdukacijaITreningList().get(i).getInstitucija().getNaziv());
-					if (institucijaDB == null) {
-						institucijaDB = cv.getEdukacijaITreningList().get(i).getInstitucija();
-						institucijaDB.setDependenciesNull();
-						institucijaDB.setMjesto(cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto());
-						institucijaRepo.save(institucijaDB);
-						cv.getEdukacijaITreningList().get(i).setInstitucija(institucijaDB);
-					}
+			}
+			for (int i = 0; i < cv.getEdukacijaITreningList().size(); i++) {
+				Institucija institucijaDB = institucijaRepo
+						.findByNaziv(cv.getEdukacijaITreningList().get(i).getInstitucija().getNaziv());
+				if (institucijaDB == null) {
+					institucijaDB = cv.getEdukacijaITreningList().get(i).getInstitucija();
+					institucijaDB.setDependenciesNull();
+					institucijaDB.setMjesto(cv.getEdukacijaITreningList().get(i).getInstitucija().getMjesto());
+					institucijaRepo.save(institucijaDB);
 					cv.getEdukacijaITreningList().get(i).setInstitucija(institucijaDB);
-
 				}
-				
-				return cv;
+				cv.getEdukacijaITreningList().get(i).setInstitucija(institucijaDB);
+
+			}
+
+			return cv;
+
+		} catch (Exception e) {
+			log.error("Error while checking for duplicates", e);
+			return null;
+		}
 	}
-	
+
+	@Transactional
+	public void CalculateDateRangeWorkEdu() {
+		eduTrenRepo.CALCULATE_DATE_DIFF_EDU();
+		radnoIRepo.CALCULATE_DATE_DIFF_WORK();
+	}
+
 	@Transactional
 	public void SaveCVtest() {
 		Date d = new Date();
@@ -372,8 +384,8 @@ public class ZivotopisDBService {
 	}
 
 	@Transactional
-	public List<Osoba> getAllCV() {
-		return Lists.newArrayList(osobaRepo.findAll());
+	public List<CV> getAllCV() {
+		return Lists.newArrayList(cvRepo.findAll());
 
 	}
 
@@ -389,8 +401,26 @@ public class ZivotopisDBService {
 	}
 
 	@Transactional
-	public void deleteCVByID(Integer id) {
-		cvRepo.deleteById(id);
+	public Boolean deleteCVByID(Integer id) {
+		try {
+			Optional<CV> cvOpt = cvRepo.findById(id);
+			if (cvOpt.get() == null) {
+
+			} else {
+				CV cv = cvOpt.get();
+				Osoba osoba = osobaRepo.findById(cv.getOsoba().getOsobaID()).get();
+				if (osoba.getZivotopisiList().size() != 1) {
+					cvRepo.deleteById(id);
+				} else if (osoba.getZivotopisiList().size() == 1) {
+					cvRepo.deleteById(id);
+					osobaRepo.deleteById(osoba.getOsobaID());
+				}
+
+			}
+			return true;
+		} catch (Exception e) {
+			return false;
+		}
 	}
 
 }
